search: create separate repo_has_file file
